strerror: add error strings for EUCLEAN and ENAVAIL
while not the only error codes presently omitted, these two are
particularly likely to be encountered in the wild.

EUCLEAN is used by linux filesystem and device drivers to report
filesystem structure corruption or data corruption.

ENAVAIL is used by some linux drivers to indicate non-availability of
a resource.

both names are new inventions to correspond to how they are actually
used, as the original kernel strings ("Structure needs cleaning" and
"No XENIX semaphores available") are not remotely meaningful or
reasonable.